Separate out Influx pump setup and Grafana setup
The influx pump and grafana need to be set up with influx user/token,
etc. This was previously part of a separate container image that is run
in parallel with everything else. This was problematic because docker
would not register at runtime the changes to the config files as these
were set up. Additionally, it was easy to get things de-synced and end
up in a situation where old tokens are configured and no longer valid.

For these reasons, we separate out the setup to run with:

$ ./compose.sh --setup-influx-test-db

This will sart up influx, grafana, and then the two setup containers,
run setup scripts, then stop everything and clean up any leftovers.
Additionally, before running, it cleans out any old config
automatically, so everything is guaranteed to be fresh.